utils: Move mtab functions to submodule
